pyproject.toml
[project]
name = "ci-agent"
version = "0.1.0"
description = "AI Agent for analyzing and optimizing GitHub CI pipelines"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.48",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
]

[project.scripts]
ci-agent = "ci_optimizer.cli:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
